+++ mmms-backend/src/main/java/com/sep/mmms_backend/exceptions/IllegalOperationException.java
@@ -1,0 +1,7 @@
+package com.sep.mmms_backend.exceptions;
+
+public class IllegalOperationException extends RuntimeException{
+    public IllegalOperationException(ExceptionMessages message){
+        super(message.toString());
+    }
+}