+++ mmms-backend/src/main/java/com/sep/mmms_backend/exceptions/InvalidRequestException.java
@@ -1,0 +1,8 @@
+package com.sep.mmms_backend.exceptions;
+
+
+public class InvalidRequestException extends RuntimeException{
+    public InvalidRequestException(ExceptionMessages ex){
+        super(ex.toString());
+    }
+}