+++ mmms-backend/src/main/java/com/sep/mmms_backend/response/ResponseMessages.java
@@ -24,6 +24,8 @@ public enum ResponseMessages {
 
     //COMMITTEE RELATED:
     COMMITTEE_CREATION_SUCCESS("Committee created successfully"),
+
+    MEMBER_CREATION_SUCCESS("Member created successfully"),
     ;